[IR] Add "transitive" unwind dest iterators
The `catchswitch` operator is an "unsplittable" EH pad; it represents
control flow which the runtime will effect outside of the compiler's
control, and so cannot be broken up or have instructions inserted along
the way to the final catchpad or outer unwind destination.  For certain
algorithms algorithms the natural way to incorporate this is to view an
invoke (or cleanupret or catchswitch) with a catchswitch unwind dest as a
multi-way branch to the successors of the catchswitch itself (and
transitively so in the case that one catchswitch unwinds to another).

Add an iterator which presents this view, which allows uniform code to
process all "real" handler(s) to which an instruction with an unwind edge
may unwind.

Also provide a variant that visits the catchswitches themselves during
iteration to accomodate callers that need to process PHIs along the way.